api/requirements.txt
Flask==2.3.2
prometheus-flask-exporter==0.23.2
psycopg2-binary==2.9.7 